Organized files and tests by functionality type
