+++ src/smartCalendarUtils.js
@@ -1,23 +1,35 @@
+// Learning TODOs
+// TODO Learn how to get caldav to work
+// TODO Learn how to print out order of lines executed - or add statements to print this out
+// TODO Learn how to debug from command line
+// TODO Figure out why babel-node does not work, expects 7.0 and I cant install
+// TODO Why does unterminated literal red squiggly appear in random place?
+
 import ICAL from 'ical.js';
 import moment from 'moment';
-import fs from 'fs';
-import readline from 'readline';
-import { google } from 'googleapis';
-
 // If modifying these scopes, delete token.json.
-const SCOPES = ['https://www.googleapis.com/auth/calendar'];
-
-const CREDENTIAL_FILE ='smart-calendar-credentials.json'
-const TOKEN_FILE = 'smart-calendar-token.json';
+export { getICalEvents, consoleLogFormatEvents };
 
 function getICalEvents(iCalData) {
     var jcal = ICAL.parse(iCalData);
     var vcal = new ICAL.Component(jcal);
     var vevents = vcal.getAllSubcomponents("vevent");
-    return vevents.map(getICalEventAtrributes);
+    return vevents.map(_getICalEventAtrributes);
 };
 
-function getICalEventAtrributes(vevent) {
+function consoleLogFormatEvents(events) {
+    if (events.length) {
+        console.log('Upcoming 10 events:');
+        events.map((event, i) => {
+            const start = event.start.dateTime || event.start.date;
+            console.log(`${start} - ${event.summary}`);
+        });
+    } else {
+        console.log('No upcoming events found.');
+    }
+}
+
+function _getICalEventAtrributes(vevent) {
     var element = {
         uid: vevent.getFirstPropertyValue("uid"),
         name: vevent.getFirstPropertyValue("summary"),
@@ -28,63 +40,3 @@ function getICalEventAtrributes(vevent) {
     return element;
 };
 
-function getCodeFromUserUsingGeneratedUrl(oAuth2Client) {
-    return new Promise(function (resolve, reject) {
-        const authUrl = oAuth2Client.generateAuthUrl({
-            access_type: 'offline',
-            scope: SCOPES,
-        });
-        console.log('Authorize this app by visiting this url:', authUrl);
-        const rl = readline.createInterface({
-            input: process.stdin,
-            output: process.stdout,
-        });
-        rl.question('Enter the code from that page here: ', (code) => {
-            rl.close();
-            resolve(code);
-        });
-    });
-}
-
-async function getOAuth2ClientFromCredentials() {
-    const unparsedCredentials = fs.readFileSync(CREDENTIAL_FILE);
-    const credentials = JSON.parse(unparsedCredentials);
-    const { client_secret, client_id, redirect_uris } = credentials.installed;
-    const oAuth2Client = new google.auth.OAuth2(
-        client_id, client_secret, redirect_uris[0]);
-    const token = await getTokenAndCreateTokenFileAndStoreTokenIfApplicable(oAuth2Client);
-    oAuth2Client.setCredentials(JSON.parse(token));
-    return oAuth2Client;
-}
-
-async function getTokenAndCreateTokenFileAndStoreTokenIfApplicable(oAuth2Client) {
-    let token = "";
-    try {
-        token = fs.readFileSync(TOKEN_FILE)
-    } catch (e) {
-         token = await getTokenAndCreateTokenFile(oAuth2Client);
-    }
-    return token;
-}
-
-
-async function getTokenAndCreateTokenFile(oAuth2Client) {
-    console.log("gocfc 1");
-    const code = await getCodeFromUserUsingGeneratedUrl(oAuth2Client);
-    return new Promise(function (resolve, reject) {
-        oAuth2Client.getToken(code, (err, token) => {
-            if (err)
-                return console.error('Error retrieving access token', err);
-            oAuth2Client.setCredentials(token);
-            // Store the token to disk for later program executions
-            fs.writeFile(TOKEN_FILE, JSON.stringify(token), (err) => {
-                if (err)
-                    console.error(err);
-                console.log('Token stored to', TOKEN_FILE);
-            });
-            resolve(JSON.stringify(token));
-        });
-    });
-}
-
-export {getICalEvents,getOAuth2ClientFromCredentials};
